split symm_basic.cpp (2/4): extract matrigen/setgroup to symm_basic_setgroup.cpp
