+++ source/source_cell/module_symmetry/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library(
     symm_other.cpp
     symm_basic.cpp
     symm_basic_order.cpp
+    symm_basic_setgroup.cpp
     symm_analysis.cpp
     symm_check.cpp
     symm_getgroup.cpp
